complete auth checks

+++ apps/nestjs-api/src/tp-jobseeker-profile-experience-records/tp-jobseeker-profile-experience-records.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common'
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common'
 import {
   TpJobseekerProfileExperienceRecordEntity,
   TpJobseekerProfileExperienceRecordEntityProps,
@@ -64,8 +68,18 @@ export class TpJobseekerProfileExperienceRecordsService {
     return await this.api.create(recordToPersist)
   }
 
-  async patch(input: TpJobseekerProfileExperienceRecordPatchInput) {
+  async patch(
+    input: TpJobseekerProfileExperienceRecordPatchInput,
+    currentUser: CurrentUserInfo
+  ) {
     const existingEntity = await this.findOne(input.id)
+
+    if (existingEntity.props.userId !== currentUser.userId) {
+      throw new UnauthorizedException(
+        'You are not authorized to update this record'
+      )
+    }
+
     const props = existingEntity.props
     const updatesSanitized = deleteUndefinedProperties(input)
     Object.entries(updatesSanitized).forEach(([key, value]) => {
@@ -76,8 +90,18 @@ export class TpJobseekerProfileExperienceRecordsService {
     await this.api.update(this.mapper.toPersistence(entityToPersist))
   }
 
-  async delete(input: TpJobseekerProfileExperienceRecordDeleteInput) {
+  async delete(
+    input: TpJobseekerProfileExperienceRecordDeleteInput,
+    currentUser: CurrentUserInfo
+  ) {
     const existingEntity = await this.findOne(input.id)
+
+    if (existingEntity.props.userId !== currentUser.userId) {
+      throw new UnauthorizedException(
+        'You are not authorized to update this record'
+      )
+    }
+
     const recordToDelete = this.mapper.toPersistence(existingEntity)
     await this.api.delete(recordToDelete)
   }
